+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## Unreleased
 ### Added
 - Support for Nyaa.si tracker.
+- Set download path for downloading torrent files.
 
 ## [1.0.3] - 2018-03-04
 ### Fixed